Allow Multiple Configurations per Widget
Use case: A shared dashboard configuration to specifiy a URL common to
all jobs exists. To use it, the widget configuration string must match
the global configuration name. Two widgets in a dashboard can load the
common URL by both specifying the same common configuration key. But
then there is no way to configure their jobs to execute at different
intervals.

Solution: Allow a widget's `config` property to be either a string or
and array. When a string is used, it points to a single configuration
key to load from the local and common configurations. When an array is
used, it points to mulitple configuration keys. Each individual key is
loaded using the same mechanism as is a single string value. Key values
specified later in the array will override matching ealier keys.

Change: The `getJobs` function in "lib/job-manager.js" is updated to
check the data type of a given `job.config`. If a `string`, it works as
normal. If an `array`, it maps the keys to a collection of each merged
local and common value. It then performs a final `extend` on the created
collection.